Added support for viewing Google documents in environment info pages
